+++ CHANGELOG.md
@@ -42,6 +42,7 @@ It is based on Tidepool Blip 1.27.
 - YLP-1594 Implement the delete button for remote monitored patients
 - YLP-1595 Plug prescription upload to backend
 - YLP-1596 Referring doctor must be saved into backend when inviting a user to remote monitoring team
+- YLP-1597 add renew dialog for a patient under remote monitoring 
 - YLP-1604 Update patient profile with new patient information
 - YLP-1607 Ask patient to consent when he joins the monitoring program
 - YLP-1609 Add button to reset patient monitoring parameters